feat: move script init to plugin
